Update 03. Triangles on Bookshelves.sql

+++ PostgreSQL/Lab Built-in Functions/03. Triangles on Bookshelves.sql
@@ -1,1 +1,1 @@
-SELECT "id", ROUND((side * height * 0.5)::numeric, 16) AS area FROM triangles;
+SELECT "id", ROUND((side * height * 0.5), 16) AS area FROM triangles;
